# packages/api-client/src/generated/types.gen.ts
// This file is auto-generated by @hey-api/openapi-ts

export type ClientOptions = {
  baseUrl: string;
};

export type AppInfoDto = {
  /**
   * Service name.
   */
  name: string;
  /**
   * Always `ok` when the API is reachable.
   */
  status: string;
};

export type LivenessDto = {
  /**
   * Always `ok` while the process serves HTTP.
   */
  status: 'ok';
};

export type DependencyCheckDto = {
  status: 'up' | 'down';
  /**
   * Round-trip time of the check.
   */
  latencyMs: number;
  /**
   * Short reason when down (`timeout`, `unreachable`); never contains hosts or credentials.
   */
  error?: string;
};

export type ReadinessChecksDto = {
  database: DependencyCheckDto;
  redis: DependencyCheckDto;
  search: DependencyCheckDto;
};

export type ReadinessDto = {
  /**
   * `ok` only if every dependency is up; otherwise HTTP 503.
   */
  status: 'ok' | 'error';
  checks: ReadinessChecksDto;
};

export type CategoryImageDto = {
  url: string;
  width: number | null;
  height: number | null;
};

export type CategoryNodeDto = {
  id: string;
  slug: string;
  /**
   * Full slug path, e.g. `zbroia/vohnepalna-zbroia`; the storefront URL is `/category/{path}`.
   */
  path: string;
  /**
   * 0 for root categories.
   */
  depth: number;
  /**
   * Name in the response locale (falls back to `uk`).
   */
  name: string;
  /**
   * Lucide icon name for menus.
   */
  icon: string | null;
  image: CategoryImageDto | null;
  /**
   * Active subcategories in menu order.
   */
  children: Array<CategoryNodeDto>;
};

export type CategoryTreeDto = {
  /**
   * Locale the names are in.
   */
  locale: string;
  /**
   * Active root categories with their active descendants.
   */
  items: Array<CategoryNodeDto>;
};

export type BreadcrumbDto = {
  name: string;
  path: string;
};

export type CategorySummaryDto = {
  id: string;
  slug: string;
  path: string;
  name: string;
  icon: string | null;
  image: CategoryImageDto | null;
};

export type CategoryDetailDto = {
  locale: string;
  id: string;
  slug: string;
  path: string;
  depth: number;
  name: string;
  description: string | null;
  icon: string | null;
  image: CategoryImageDto | null;
  /**
   * Root → this category (inclusive).
   */
  breadcrumbs: Array<BreadcrumbDto>;
  /**
   * Direct active subcategories.
   */
  children: Array<CategorySummaryDto>;
};

export type ErrorDetailDto = {
  /**
   * Dot path of the invalid field, e.g. `address.city` or `items.0.qty`.
   */
  field: string;
  /**
   * Human-readable problems with this field.
   */
  messages: Array<string>;
};

export type ErrorResponseDto = {
  /**
   * HTTP status code.
   */
  statusCode: number;
  /**
   * Machine-readable code, e.g. `VALIDATION_FAILED`, `NOT_FOUND`.
   */
  code: string;
  /**
   * Human-readable summary; safe to show, never contains internals.
   */
  message: string;
  /**
   * Correlates with server logs and the `X-Request-Id` header.
   */
  requestId: string;
  /**
   * Per-field problems (validation errors).
   */
  details?: Array<ErrorDetailDto>;
};

export type AppGetRootData = {
  body?: never;
  path?: never;
  query?: never;
  url: '/api/v1';
};

export type AppGetRootErrors = {
  /**
   * Error (unified format, see `code`)
   */
  default: ErrorResponseDto;
};

export type AppGetRootError = AppGetRootErrors[keyof AppGetRootErrors];

export type AppGetRootResponses = {
  200: AppInfoDto;
};

export type AppGetRootResponse = AppGetRootResponses[keyof AppGetRootResponses];

export type HealthLivenessData = {
  body?: never;
  path?: never;
  query?: never;
  url: '/health';
};

export type HealthLivenessErrors = {
  /**
   * Error (unified format, see `code`)
   */
  default: ErrorResponseDto;
};

export type HealthLivenessError = HealthLivenessErrors[keyof HealthLivenessErrors];

export type HealthLivenessResponses = {
  200: LivenessDto;
};

export type HealthLivenessResponse = HealthLivenessResponses[keyof HealthLivenessResponses];

export type HealthReadinessData = {
  body?: never;
  path?: never;
  query?: never;
  url: '/health/ready';
};

export type HealthReadinessErrors = {
  /**
   * A dependency is down
   */
  503: ReadinessDto;
  /**
   * Error (unified format, see `code`)
   */
  default: ErrorResponseDto;
};

export type HealthReadinessError = HealthReadinessErrors[keyof HealthReadinessErrors];

export type HealthReadinessResponses = {
  200: ReadinessDto;
};

export type HealthReadinessResponse = HealthReadinessResponses[keyof HealthReadinessResponses];

export type CategoriesGetTreeData = {
  body?: never;
  headers?: {
    /**
     * Used when `locale` is absent
     */
    'Accept-Language'?: string;
  };
  path?: never;
  query?: {
    /**
     * Response language (`uk`, `ru`, `en`). Unsupported → `Accept-Language` → `uk`.
     */
    locale?: string;
  };
  url: '/api/v1/categories';
};

export type CategoriesGetTreeErrors = {
  /**
   * Error (unified format, see `code`)
   */
  default: ErrorResponseDto;
};

export type CategoriesGetTreeError = CategoriesGetTreeErrors[keyof CategoriesGetTreeErrors];

export type CategoriesGetTreeResponses = {
  200: CategoryTreeDto;
};

export type CategoriesGetTreeResponse =
  CategoriesGetTreeResponses[keyof CategoriesGetTreeResponses];

export type CategoriesGetByPathData = {
  body?: never;
  headers?: {
    /**
     * Used when `locale` is absent
     */
    'Accept-Language'?: string;
  };
  path?: never;
  query: {
    /**
     * Response language (`uk`, `ru`, `en`). Unsupported → `Accept-Language` → `uk`.
     */
    locale?: string;
    /**
     * Slug path without leading/trailing slashes.
     */
    path: string;
  };
  url: '/api/v1/categories/by-path';
};

export type CategoriesGetByPathErrors = {
  /**
   * Error (unified format, see `code`)
   */
  default: ErrorResponseDto;
};

export type CategoriesGetByPathError = CategoriesGetByPathErrors[keyof CategoriesGetByPathErrors];

export type CategoriesGetByPathResponses = {
  200: CategoryDetailDto;
};

export type CategoriesGetByPathResponse =
  CategoriesGetByPathResponses[keyof CategoriesGetByPathResponses];